Update preact to get useIntl() support
We might want to make use of useIntl() in implementing
localization for subject, message, and selection of
recipients.

Note that `preact-compat` has been merged into `preact`,
and includes should refer to `preact/compat` instead of
`preact-compat`.

As part of this update, remove react-intl/locale-data/*
includes, which have been migrated to a different package.
These includes don't appear to be necessary.